requirements.txt
Babel==2.5.3
beautifulsoup4==4.6.0
certifi==2018.1.18
chardet==3.0.4
debtcollector==1.19.0
Django==2.0.1
django-bootstrap3==9.1.0
enum-compat==0.0.2
eventlet==0.20.0
greenlet==0.4.12
idna==2.6
iso8601==0.1.12
lxml==4.1.1
monotonic==1.4
netaddr==0.7.19
oslo.i18n==3.19.0
pbr==3.1.1
Pillow==5.0.0
pyparsing==2.2.0
PyPDF2==1.26.0
pytz==2017.3
requests==2.18.4
six==1.11.0
urllib3==1.22
wrapt==1.10.11
